Getting a list of recipes to display.
