+++ RecipeUIClassLib/Areas/Recipes/Pages/Index.cshtml
@@ -1,13 +1,34 @@
-﻿@page
-@model RecipeUIClassLib.Areas.Recipes.Pages.IndexModel
+@page
+@model RecipeUIClassLib.Areas.Recipes.Models.IndexModel
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 
-<h4>@ViewData["Title"]</h4>
-<partial name="_Message" />
-<p> From  RecipeUIClassLib\Areas\Recipes\Pages\Index.cshtml</p>
-<ul>
-    @foreach (var recipe in Model.Recipes)
-    {
-        <li>@recipe.Title</li>
-    }
-</ul>
+<!-- <div class="album py-5 bg-light"> -->
+    <div class="container">
+        <div class="row">
+            @foreach (var recipe in Model.Recipes)
+            {
+                <div class="col-md-4">
+                    <div class="card mb-4 shadow-sm">
+                        <svg class="bd-placeholder-img card-img-top" width="100%" height="128" xmlns="http://www.w3.org/2000/svg"
+                            preserveAspectRatio="xMidYMid slice" focusable="false" role="img" aria-label="Placeholder: No Picture">
+                            <title>Placeholder</title>
+                            <rect width="33%" height="33%" fill="#55595c"/>
+                            <text x="30%" y="50%" fill="#eceeef" dy=".3em">No Picture</text>
+                        </svg>
+                        <div class="card-body">
+                            <h5 class="card-title">@recipe.Title</h5>
+                            <p class="card-text">@recipe.Description</p>
+                            <div class="d-flex justify-content-between align-items-center">
+                                <div class="btn-group">
+                                <button type="button" class="btn btn-sm btn-outline">View</button>
+                                <button type="button" class="btn btn-sm btn-outline-secondary">Edit</button>
+                                </div>
+                                <small class="text-muted">9 mins</small>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            }
+        </div>
+    </div>
+<!-- </div> -->
